[refactor] investment comment 내용 refactoring

+++ src/main/java/com/finlearn/simulationservice/application/investment/dto/response/BuyStockResponse.java
@@ -1,6 +1,5 @@
 package com.finlearn.simulationservice.application.investment.dto.response;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record BuyStockResponse(
@@ -9,8 +8,8 @@ public record BuyStockResponse(
         String stockName,
         String tradeType,
         long quantity,
-        BigDecimal price,
-        BigDecimal totalAmount,
-        BigDecimal cashBalanceAfterTrade
+        long price,
+        long totalAmount,
+        long cashBalanceAfterTrade
 ) {
 }
