+++ src/main/java/com/finlearn/simulationservice/application/investment/dto/response/SellStockResponse.java
@@ -1,14 +1,12 @@
 package com.finlearn.simulationservice.application.investment.dto.response;
 
-import java.math.BigDecimal;
-
 public record SellStockResponse(
         String stockCode,
         String stockName,
         long sellQuantity,
-        BigDecimal sellPrice,
-        BigDecimal totalSellAmount,
+        long sellPrice,
+        long totalSellAmount,
         long remainingQuantity,
-        BigDecimal cashBalance
+        long cashBalance
 ) {
 }
